Struct rendy_frame::cirque::Cirque

pub struct Cirque<T, I = T, P = T> { /* fields omitted */ }

Resource cirque. It simplifies using multiple resources when same resource cannot be used simulteneously.

Implementations

impl<B, C, P, L> Cirque<CommandBuffer<B, C, ExecutableState<MultiShot<NoSimultaneousUse>, P>, L, IndividualReset>, CommandBuffer<B, C, InitialState, L, IndividualReset>, CommandBuffer<B, C, PendingState<ExecutableState<MultiShot<NoSimultaneousUse>, P>>, L, IndividualReset>> where
B: Backend,
L: Level,
C: Capability,
P: RenderPassRelation<L>,

pub fn encode<'a>(
&'a mut self,
frames: &Frames<B>,
pool: &mut CommandPool<B, C, IndividualReset>,
encode: impl FnOnce(CommandCirqueRef<'a, B, C, P, L>) -> CommandReadyRef<'a, B, C, P, L>
) -> Submit<B, NoSimultaneousUse, L, P>

Encode and submit.

impl<T, I, P> Cirque<T, I, P>

pub fn new() -> Self

Create new empty Cirque

pub fn dispose(self, dispose: impl FnMut(Either<T, P>))

Dispose of the Cirque.

pub fn get<B: Backend>(
&mut self,
frames: &Frames<B>,
alloc: impl FnOnce() -> I,
complete: impl Fn(P) -> T
) -> CirqueRef<T, I, P>

Get CirqueRef for specified frames range. Allocate new instance in initial state if no ready values exist.